[Impl] Switch to seperate dataspace

+++ vertigo-orchestra-webapi/src/main/java/io/vertigo/orchestra/webapi/services/ExecutionServices.java
@@ -21,7 +21,7 @@ public interface ExecutionServices extends StoreServices {
 
 	OActivityExecutionUi getActivityExecutionById(Long aceId);
 
-	DtList<OProcessExecutionUi> getProcessExecutionsByProId(Long proId, String status, Long limit, Long offset);
+	DtList<OProcessExecutionUi> getProcessExecutionsByProId(Long proId, String status, Integer limit, Integer offset);
 
 	DtList<OActivityExecutionUi> getActivityExecutionsByPreId(Long preId);
